seperated checkoutStore with attestation store
